Make basic suite execute
