--- packages/upgrade/test/__fixtures__/hooks/basic.spec.ts
@@ -1,6 +1,0 @@
-import { expect } from 'vitest';
-import type { SpecRun } from '../../specDef';
-
-export const run: SpecRun = async (Component) => {
-  expect(Component).toBeDefined();
-};
