--- packages/upgrade/test/__fixtures__/hooks/basic.trpc.ts
@@ -1,5 +1,0 @@
-import { appRouter, testReactResource } from '../../__helpers';
-
-export const ctx = testReactResource(appRouter);
-export const trpc = ctx.rq;
-export const useTRPC = ctx.trq.useTRPC;
