+++ packages/upgrade/test/specDef.ts
@@ -1,3 +1,3 @@
-export type SpecRun = (Component: React.Component) => Promise<void>;
+export type SpecRun = (Component: React.FC) => Promise<void>;
 export type SpecDefFile = { run: SpecRun };
-export type ComponentFile = { Component: React.Component };
+export type ComponentFile = { Component: React.FC };
