+++ packages/upgrade/test/transforms.test.ts
@@ -27,49 +27,55 @@ const snapshotTestTransform = async (file: string, transform: any) => {
   );
   const formatted = await formatFile(file, transformed);
 
-  await expect(formatted).toMatchFileSnapshot(file.replace('.ts', '.snap.ts'));
+  await expect(formatted).toMatchFileSnapshot(
+    file.replace('.tsx', '.snap.tsx'),
+  );
 };
 
 function isFixture(file: string) {
   return (
-    !file.endsWith('.snap.ts') &&
-    !file.endsWith('spec.ts') &&
-    !file.endsWith('.trpc.ts')
+    !file.endsWith('.snap.tsx') &&
+    !file.endsWith('.spec.tsx') &&
+    !file.endsWith('.trpc.tsx')
   );
+}
+
+async function executeTests(fixturesDir: string, file: string) {
+  const fixtureFile = join(fixturesDir, file);
+
+  const specFile = fixtureFile.replace('.tsx', '.spec.tsx');
+  if (!existsSync(specFile)) {
+    expect.fail(`There is no .spec.ts file at ${specFile}`);
+  }
+
+  const spec = (await import(specFile)) as SpecDefFile;
+
+  const fixture = (await import(fixtureFile)) as ComponentFile;
+  await spec.run(fixture.Component);
+
+  await snapshotTestTransform(fixtureFile, hooksTransform);
+
+  const snapshotFile = join(fixturesDir, file.replace('.tsx', '.snap.tsx'));
+  const snapshot = (await import(snapshotFile)) as ComponentFile;
+  await spec.run(snapshot.Component);
 }
 
 describe('hooks', () => {
   const literal = './__fixtures__/hooks'; // idk why but Vite seems to do some shit when the string is in-lined to URL
   const fixturesDir = new URL(literal, import.meta.url).pathname;
 
-  const ONLY_RUN: string[] = ['basic.ts'];
+  const ONLY_RUN: string[] = [];
 
   const fixtures = ONLY_RUN.length
     ? ONLY_RUN
     : readdirSync(fixturesDir).filter(isFixture);
 
   it.each(fixtures)('hooks %s', async (file) => {
-    const fixtureFile = join(fixturesDir, file);
-
-    const specFile = fixtureFile.replace('.ts', '.spec.ts');
-    if (!existsSync(specFile)) {
-      expect.fail(`There is no .spec.ts file at ${specFile}`);
-    }
-
-    const spec = (await import(specFile)) as SpecDefFile;
-
-    const fixture = (await import(fixtureFile)) as ComponentFile;
-    await spec.run(fixture.Component);
-
-    await snapshotTestTransform(fixtureFile, hooksTransform);
-
-    const snapshotFile = join(fixturesDir, file.replace('.ts', '.snap.ts'));
-    const snapshot = (await import(snapshotFile)) as ComponentFile;
-    await spec.run(snapshot.Component);
+    await executeTests(fixturesDir, file);
   });
 });
 
-describe.skip('provider', () => {
+describe('provider', () => {
   const literal = './__fixtures__/provider'; // idk why but Vite seems to do some shit when the string is in-lined to URL
   const fixturesDir = new URL(literal, import.meta.url).pathname;
 
@@ -80,12 +86,6 @@ describe.skip('provider', () => {
     : readdirSync(fixturesDir).filter(isFixture);
 
   it.each(fixtures)('provider %s', async (file) => {
-    const fixtureFile = join(fixturesDir, file);
-    await snapshotTestTransform(fixtureFile, providerTransform);
-
-    const specFile = fixtureFile.replace('.ts', '.spec.ts');
-    if (!existsSync(specFile)) {
-      expect.fail(`There is no .spec.ts file at ${specFile}`);
-    }
+    await executeTests(fixturesDir, file);
   });
 });
