+++ packages/upgrade/vitest.config.ts
@@ -4,5 +4,6 @@ import baseConfig from '../../vitest.config';
 export default mergeConfig(baseConfig, {
   test: {
     exclude: ['test/__fixtures__/*/*.spec.ts', 'node_modules'],
+    setupFiles: ['./test/setupTests.ts'],
   },
 });
